+++ libs/User.php
@@ -13,6 +13,7 @@ namespace Mertowitch\Phpneeds
 {
 
     use PDO;
+    use PDOException;
 
     /**
      * Class User
@@ -22,11 +23,19 @@ namespace Mertowitch\Phpneeds
 
         private Database $objDatabase;
 
+        /**
+         * @param Database $objDatabase
+         */
         public function __construct( Database $objDatabase )
         {
             $this->objDatabase = $objDatabase;
         }
 
+        /**
+         * @param int $userID
+         *
+         * @return array|bool
+         */
         public function getByID( int $userID ): array|bool
         {
             if ( $return = $this->_getByID( $userID ) )
@@ -37,6 +46,11 @@ namespace Mertowitch\Phpneeds
             return false;
         }
 
+        /**
+         * @param int $userID
+         *
+         * @return array|bool
+         */
         private function _getByID( int $userID ): array|bool
         {
             $qryUser = $this->objDatabase->prepare( 'SELECT * FROM `users` WHERE `ID`=:ID LIMIT 1' );
@@ -46,6 +60,9 @@ namespace Mertowitch\Phpneeds
             return $qryUser->fetch();
         }
 
+        /**
+         * @return array|bool
+         */
         public function getAll(): array|bool
         {
             if ( $return = $this->_getAll() )
@@ -56,6 +73,9 @@ namespace Mertowitch\Phpneeds
             return false;
         }
 
+        /**
+         * @return array|bool
+         */
         private function _getAll(): array|bool
         {
             $qryUser = $this->objDatabase->query( 'SELECT * FROM `users`' );
@@ -63,6 +83,11 @@ namespace Mertowitch\Phpneeds
             return $qryUser->fetchAll();
         }
 
+        /**
+         * @param string $username
+         *
+         * @return array|bool
+         */
         public function getByUsername( string $username ): array|bool
         {
             if ( $return = $this->_getByUsername( $username ) )
@@ -73,6 +98,11 @@ namespace Mertowitch\Phpneeds
             return false;
         }
 
+        /**
+         * @param string $username
+         *
+         * @return array|bool
+         */
         private function _getByUsername( string $username ): array|bool
         {
             $qryUser = $this->objDatabase->prepare( 'SELECT * FROM `users` WHERE `username`=:username LIMIT 1' );
@@ -82,6 +112,12 @@ namespace Mertowitch\Phpneeds
             return $qryUser->fetch();
         }
 
+        /**
+         * @param string $username
+         * @param string $password
+         *
+         * @return bool
+         */
         public function login( string $username, string $password ): bool
         {
             if ( ! $this->_verifyCredential( $username, $password ) )
@@ -92,9 +128,51 @@ namespace Mertowitch\Phpneeds
             return true;
         }
 
+        /**
+         * Verifying credential for given username and password
+         *
+         * @param string $username
+         * @param string $password
+         *
+         * @return bool
+         */
         private function _verifyCredential( string $username, string $password ): bool
         {
             return ( $arrUser = $this->_getByUsername( $username ) ) && $arrUser['password'] === md5( $password );
+        }
+
+        /**
+         * @param string $username
+         * @param string $password
+         *
+         * @return bool
+         */
+        public function createUser( string $username, string $password ): bool
+        {
+            $result = false;
+
+            try
+            {
+                $schema = $this->objDatabase->getSchema( 'USER' );
+
+                $sqlNewUser = "INSERT INTO `{$schema->NAME}` ( `{$schema->FIELDS['USERNAME']['NAME']}`, `{$schema->FIELDS['PASSWORD']['NAME']}` ) VALUES(:username, :password)";
+
+                $qryNewUser = $this->objDatabase->prepare( $sqlNewUser );
+
+                $qryNewUser->bindParam( ':username', $username, PDO::PARAM_STR );
+                $qryNewUser->bindParam( ':password', $password, PDO::PARAM_STR );
+
+                if ( $qryNewUser->execute() )
+                {
+                    $result = true;
+                }
+            }
+            catch ( PDOException $e )
+            {
+                return false;
+            }
+
+            return $result;
         }
 
     }