+++ CHANGELOG.md
@@ -4,6 +4,16 @@
 ### Added
 - [**User class**](/libs/User.php) added
 - "PREFIX" added to [Redis class](/libs/Redis.php) and [Redis config](/confs/conf.redis.php)
+- "/tools/database/**create_table_users.php**" added
+- new methods added to [Database class](/libs/Database.php)
+  - ->createTable( string $tableName ): bool
+  - ->isTableExist( string $tableName ): bool
+  - ->getSchema( string $tableName ): object
+- new methods added to **User class**
+  - ->login( string $username, string $password ): bool
+  - ->_verifyCredential( string $username, string $password ): bool
+  - ->createUser( string $username, string $password ): bool
+
 ### Changed
 - **User object** added to [init.php](/common/init.php)
 - **/images** directory moved to [/public/assets/images](/public/assets/images)
@@ -14,9 +24,9 @@
 
 ## [v0.3.2](https://github.com/mertowitch/PHPneeds/compare/v0.3.1...v0.3.2) - 2021-09-20
 ### Changed
-- README.md updated ([detail](https://github.com/mertowitch/PHPneeds/compare/v0.3.1...v0.3.2#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5))
-- Database class changed ([detail](https://github.com/mertowitch/PHPneeds/compare/v0.3.1...v0.3.2#diff-a71fd8e5cd2d33de5278493e2f5f04aeea25ed26cf500d65742d492eb39afec8))
-- Session class changed ([detail](https://github.com/mertowitch/PHPneeds/compare/v0.3.1...v0.3.2#diff-e6e35f5873800dc7a0e2c9ce9e96e63f488fa446e1019f59797b9c1f08575dcc))
+- README.md updated ([detail](https://github.com/mertowitch/PHPneeds/compare/v0.3.1...v0.3.2))
+- Database class changed ([detail](https://github.com/mertowitch/PHPneeds/compare/v0.3.1...v0.3.2))
+- Session class changed ([detail](https://github.com/mertowitch/PHPneeds/compare/v0.3.1...v0.3.2))
 
 ## [v0.3.1](https://github.com/mertowitch/PHPneeds/compare/v0.3.0...v0.3.1) - 2021-09-17
 ### Changed